pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "email-triage-env"
version = "0.1.0"
description = "OpenEnv Email Triage Environment - an agentic email management environment"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.3",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.0.0",
    "fastmcp>=2.0.0",
    "httpx>=0.27.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["server", "models"]

[tool.hatch.metadata]
allow-direct-references = true

[project.scripts]
server = "server.app:main"
